pyproject.toml
[project]
name = "sysadmin-agents"
version = "0.1.0"
description = "Enterprise-grade AI agents for Linux/RHEL system administration using linux-mcp-server"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
    { name = "Sysadmin Agents Contributors" }
]
keywords = ["linux", "rhel", "sysadmin", "mcp", "agents", "adk", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "google-adk>=1.18.0",
    "linux-mcp-server>=0.1.0a0",  # Accept alpha versions
    "pyyaml>=6.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
# Multi-model support (GPT-4, Claude, etc.)
multimodel = [
    "litellm>=1.40.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "litellm>=1.40.0",  # Include for testing multi-model
]

# [project.urls]
# Update these URLs when you publish to a real repository
# Homepage = "https://github.com/your-org/sysadmin-agents"
# Repository = "https://github.com/your-org/sysadmin-agents.git"
# Issues = "https://github.com/your-org/sysadmin-agents/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "core"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

